split propertyTree redux into properties and owners
