+++ src/api/hooks.ts
@@ -8,12 +8,12 @@ export function useOpenSpaceApi() {
 }
 
 export function useGetProperty(uri: string): Property | undefined {
-  return useAppSelector((state) => state.propertyTree.props.properties[uri]);
+  return useAppSelector((state) => state.properties.properties[uri]);
 }
 
 function useGetPropertyValue<T>(uri: string, propertyType: string): T | undefined {
   return useAppSelector((state) => {
-    const prop = state.propertyTree.props.properties[uri];
+    const prop = state.properties.properties[uri];
     if (prop && prop?.description.type === propertyType) {
       throw Error(`Requested a ${propertyType} but got a ${prop.description.type}`);
     }
