+++ src/redux/events/eventsMiddleware.ts
@@ -4,8 +4,9 @@ import { Topic } from 'openspace-api-js';
 import { api } from '@/api/api';
 import { onCloseConnection, onOpenConnection } from '@/redux/connection/connectionSlice';
 import { AppStartListening } from '@/redux/listenerMiddleware';
+import { removeProperties } from '@/redux/propertytree/properties/propertiesSlice';
+import { removePropertyOwners } from '@/redux/propertytree/propertyowner/propertyOwnerSlice';
 import { addUriToPropertyTree } from '@/redux/propertytree/propertyTreeMiddleware';
-import { removePropertyOwners } from '@/redux/propertytree/propertyTreeSlice';
 import { EventData } from '@/types/event-types';
 
 let eventTopic: Topic;
@@ -19,6 +20,7 @@ function handleData(dispatch: Dispatch<UnknownAction>, data: EventData) {
       break;
     case 'PropertyTreePruned':
       dispatch(removePropertyOwners({ uris: [data.Uri] }));
+      dispatch(removeProperties({ uris: [data.Uri] }));
       //   dispatch(refreshGroups) // TODO: add refresh groups slice action
       break;
     default:
