+++ src/redux/listenerMiddleware.ts
@@ -5,6 +5,7 @@ import { addConnectionListener } from './connection/connectionMiddleware';
 import { addEventsListener } from './events/eventsMiddleware';
 import { addExoplanetListener } from './exoplanets/exoplanetsMiddleware';
 import { addFlightControllerListener } from './flightcontroller/flightControllerMiddleware';
+import { addPropertiesListener } from './propertytree/properties/propertiesMiddleware';
 import { addPropertyTreeListener } from './propertytree/propertyTreeMiddleware';
 import { addSessionRecordingListener } from './sessionrecording/sessionRecordingMiddleware';
 import { addTimeListener } from './time/timeMiddleware';
@@ -20,6 +21,7 @@ addSessionRecordingListener(startAppListening);
 addActionsListener(startAppListening);
 addExoplanetListener(startAppListening);
 addEventsListener(startAppListening);
+addPropertiesListener(startAppListening);
 addPropertyTreeListener(startAppListening);
 addFlightControllerListener(startAppListening);
 addTimeListener(startAppListening);
